add line chart

+++ app/src/main/java/wilderpereira/com/feelingfy/main/PresentationItem.kt
@@ -11,7 +11,7 @@ import java.io.File
 @Parcelize
 data class PresentationItem(
         var title: String?,
-        val quality: Double?,
+        val quality: Float?,
         val image: File?,
         val picture: Picture?
 ) : Parcelable
